requirements/common.txt
# Core Stuff
# -------------------------------------
Django==2.0.2

# Configuration
# -------------------------------------
argon2-cffi==18.1.0
django-environ==0.4.4
django-sites==0.10
python-dotenv==0.7.1



# Extensions
# -------------------------------------
pytz==2018.3

# Models
# -------------------------------------
psycopg2-binary==2.7.4

Pillow==5.0.0
django-extensions==1.9.9
django-uuid-upload-path==1.0.0
django-versatileimagefield==1.9

# REST APIs
# -------------------------------------
djangorestframework==3.7.7
django-rest-swagger==2.1.2

# LOGGING
# -------------------------------------
django-log-request-id==1.3.2

# Documentation
# -------------------------------------
mkdocs==0.17.2

# Raven is the Sentry client
# --------------------------
raven==6.5.0

# Auth Stuff
# -------------------------------------
PyJWT==1.5.3
django-mail-templated==2.6.5

# WEBPACK
# -------------------------------------
django-webpack-loader==0.3.3
